// src/app/daily-mission/daily-mission.interface.ts
// Generated by https://quicktype.io
export interface DailyMissionRecord {
  finishedDate: string | Date;
  missionType: string;
  missionTypeCht: string;
  iconCode: string;
  subject: string;
  point: number;
  description: string;
}

export interface FinishedMission {
  mental: MissionItem[];
  skill: MissionItem[];
  exercise: MissionItem[];
  social: MissionItem[];
  entertainment: MissionItem[];
}

export interface MissionItem {
  titel: string;
  point: number;
}

export interface RadarChartData {
  data: number[];
  label: string;
}

export interface MonthSelectOption {
  month: number;
  text: string;
}